// Frontend/src/types/generated/models/AddressCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AddressCreate = {
    country: string;
    city: string;
    stateOrProvince: string | null;
    street: string;
    postalCode: string;
    longitude: number | null;
    latitude: number | null;
};